# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForSEII2526.API",
    "description": "API for renting and purchasing devices",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Purchases/GetPurchase": {
      "get": {
        "tags": [
          "Purchases"
        ],
        "operationId": "GetPurchase",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Purchases/CreatePurchase": {
      "post": {
        "tags": [
          "Purchases"
        ],
        "operationId": "CreatePurchase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseForCreateDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseForCreateDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PurchaseForCreateDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDetailDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PaymentMethodType": {
        "enum": [
          "CreditCard",
          "PayPal",
          "Cash"
        ],
        "type": "string"
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "PurchaseDetailDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date-time"
          },
          "customerFullName": {
            "type": "string",
            "nullable": true
          },
          "deliveryAddress": {
            "type": "string",
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "totalQuantity": {
            "type": "integer",
            "format": "int32"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItemDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "PurchaseForCreateDTO": {
        "required": [
          "customerFirstName",
          "customerLastName",
          "customerUserName",
          "deliveryAddress",
          "paymentMethod"
        ],
        "type": "object",
        "properties": {
          "customerUserName": {
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "customerFirstName": {
            "maxLength": 30,
            "minLength": 2,
            "type": "string"
          },
          "customerLastName": {
            "maxLength": 30,
            "minLength": 2,
            "type": "string"
          },
          "deliveryAddress": {
            "maxLength": 80,
            "minLength": 10,
            "type": "string",
            "format": "multiline"
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethodType"
          },
          "purchaseItems": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItemDTO"
            },
            "nullable": true
          },
          "TotalQuantity": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "PurchaseItemDTO": {
        "type": "object",
        "properties": {
          "deviceID": {
            "type": "integer",
            "format": "int32"
          },
          "quantity": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "model": {
            "type": "string",
            "nullable": true
          },
          "color": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}